+++ app/src/main/java/com/codesoom/assignment/controllers/TaskController.java
@@ -1,44 +1,43 @@
 // TODO
-// 1. Read Collection - GET /tasks => work in progress..
+// 1. Read Collection - GET /tasks => 완료
 // 2. Read Item - GET /tasks/{id}
-// 3. Create - POST /tasks
+// 3. Create - POST /tasks => 완료
 // 4. Update - PUT/PATCH /tasks/{id}
 // 5. Delete - DELETE /tasks/{id}
 
 package com.codesoom.assignment.controllers;
 
 import com.codesoom.assignment.models.Task;
+import com.codesoom.assignment.repositories.TaskRepository;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import java.util.ArrayList;
 import java.util.List;
 
 @RestController
 @RequestMapping("/tasks")
 public class TaskController {
-    private List<Task> tasks = new ArrayList<>();
-    private Long newId = 0L;
+    private TaskRepository taskRepository;
+
+    public TaskController(TaskRepository taskRepository) {
+        this.taskRepository = taskRepository;
+    }
 
     @GetMapping
     public List<Task> list() {
-        return tasks;
+        return taskRepository.findAll();
     }
 
     @PostMapping
     public Task create(@RequestBody Task task) {
-        task.setId(generateId());
-        tasks.add(task);
+        // TODO : validation error...
+//        if (task.getTitle().isBlank()) {
+//        }
+        taskRepository.add(task);
         
         return task;
     }
-
-    private Long generateId() {
-        newId += 1;
-        return newId;
-    }
-
 }
